clk_pll_deskew.c: 7: yes: MISRA.PTR.ARITH.2012
Rule 18.4, pointer arithmetic violations in register address
calculations. Fixed 13 violations by adding explicit uintptr_t
casts to make pointer arithmetic type-safe and compliant.
IDs: 820-832. TFA enforced - pointer arithmetic must be explicit
for hardware register access safety.

+++ plat/ti/k3low/common/drivers/pm/clock/clk_pll_deskew.c
@@ -129,7 +129,7 @@ static bool clk_pll_deskew_check_lock(struct clk *clock_ptr)
 	pll = container_of(data_pll, const struct clk_data_pll_deskew,
 			   data_pll);
 
-	stat = (uint32_t)readl(pll->base + PLL_DESKEW_STAT(pll->idx));
+	stat = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_STAT(pll->idx));
 	return (stat & PLL_DESKEW_STAT_LOCK) != 0U;
 }
 
@@ -175,7 +175,7 @@ static bool clk_pll_deskew_is_bypass(struct clk *clock_ptr)
 			   data_pll);
 
 	/* IDLE Bypass */
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 	ret = (ctrl & PLL_DESKEW_CTRL_BYPASS_EN) != 0U;
 
 	return ret;
@@ -194,7 +194,7 @@ static int32_t clk_pll_deskew_bypass(struct clk *clock_ptr, bool bypass)
 	pll = container_of(data_pll, const struct clk_data_pll_deskew,
 			   data_pll);
 
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 	if (bypass) {
 		/* Enable bypass */
 		ctrl |= PLL_DESKEW_CTRL_BYPASS_EN;
@@ -202,7 +202,7 @@ static int32_t clk_pll_deskew_bypass(struct clk *clock_ptr, bool bypass)
 		/* Disable bypass */
 		ctrl &= ~PLL_DESKEW_CTRL_BYPASS_EN;
 	}
-	ti_clk_writel(ctrl, pll->base + PLL_DESKEW_CTRL(pll->idx));
+	ti_clk_writel(ctrl, (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 	return SUCCESS;
 }
 
@@ -232,8 +232,8 @@ static bool clk_pll_deskew_program_freq(struct clk *clock_ptr,
 	}
 
 	/* Program the new rate */
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
-	div_ctrl = (uint64_t)readl(pll->base + PLL_DESKEW_DIV_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
+	div_ctrl = (uint64_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_DIV_CTRL(pll->idx));
 
 	div_ctrl &= ~PLL_DESKEW_DIV_CTRL_FB_DIV_MASK;
 	switch (pllm_val_p) {
@@ -285,10 +285,10 @@ static bool clk_pll_deskew_program_freq(struct clk *clock_ptr,
 	div_ctrl |= i << PLL_DESKEW_DIV_CTRL_POST_DIV_SHIFT;
 
 	if (ret) {
-		ti_clk_writel(ctrl, pll->base + PLL_DESKEW_CTRL(pll->idx));
+		ti_clk_writel(ctrl, (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 	}
 
-	ti_clk_writel((uint32_t) div_ctrl, pll->base + PLL_DESKEW_DIV_CTRL(pll->idx));
+	ti_clk_writel((uint32_t) div_ctrl, (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_DIV_CTRL(pll->idx));
 
 	if (ret && (clock_ptr->ref_count != 0U)) {
 		/* Take the PLL out of bypass */
@@ -334,8 +334,8 @@ static uint32_t clk_pll_deskew_set_freq(struct clk *clock_ptr,
 	pll = container_of(data_pll, const struct clk_data_pll_deskew,
 			   data_pll);
 
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
-	div_ctrl = (uint64_t)readl(pll->base + PLL_DESKEW_DIV_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
+	div_ctrl = (uint64_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_DIV_CTRL(pll->idx));
 
 	/* Check current values */
 	prev_pllm = (uint64_t) div_ctrl & PLL_DESKEW_DIV_CTRL_FB_DIV_MASK;
@@ -412,8 +412,8 @@ static uint32_t clk_pll_deskew_get_freq_internal(struct clk *clock_ptr)
 
 	data_pll = container_of(clock_data->data, const struct clk_data_pll, data);
 	pll = container_of(data_pll, const struct clk_data_pll_deskew, data_pll);
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
-	div_ctrl = (uint64_t)readl(pll->base + PLL_DESKEW_DIV_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
+	div_ctrl = (uint64_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_DIV_CTRL(pll->idx));
 
 	pllm = (uint64_t) div_ctrl & PLL_DESKEW_DIV_CTRL_FB_DIV_MASK;
 	pllm >>= PLL_DESKEW_DIV_CTRL_FB_DIV_SHIFT;
@@ -604,16 +604,16 @@ static int32_t clk_pll_deskew_init_internal(struct clk *clock_ptr)
 	 * written value.
 	 */
 	ti_clk_writel(PLL_DESKEW_LOCKKEY0_VALUE,
-		      pll->base + PLL_DESKEW_LOCKKEY0(pll->idx));
+		      (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_LOCKKEY0(pll->idx));
 	ti_clk_writel(PLL_DESKEW_LOCKKEY1_VALUE,
-		      pll->base + PLL_DESKEW_LOCKKEY1(pll->idx));
+		      (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_LOCKKEY1(pll->idx));
 
-	ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_CTRL(pll->idx));
+	ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 
 	/* Make sure PLL is enabled */
 	if ((ctrl & PLL_DESKEW_CTRL_PD_EN) == 0U) {
 		ctrl |= PLL_DESKEW_CTRL_PD_EN;
-		ti_clk_writel(ctrl, pll->base + PLL_DESKEW_CTRL(pll->idx));
+		ti_clk_writel(ctrl, (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 		osal_delay(1U); /* Wait 1us */
 	}
 
@@ -626,17 +626,17 @@ static int32_t clk_pll_deskew_init_internal(struct clk *clock_ptr)
 		ctrl &= ~PLL_DESKEW_CTRL_INTL_BYP_EN;
 	}
 
-	ti_clk_writel(ctrl, pll->base + PLL_DESKEW_CTRL(pll->idx));
+	ti_clk_writel(ctrl, (uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CTRL(pll->idx));
 
 	/* Enable all HSDIV outputs */
-	cfg = (uint32_t)readl(pll->base + PLL_DESKEW_CFG(pll->idx));
+	cfg = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_CFG(pll->idx));
 	for (i = 0U; (i < 16U); i++) {
 		/* Enable HSDIV output if present */
 		if (((1UL << (i + 16UL)) & cfg) != 0UL) {
-			ctrl = (uint32_t)readl(pll->base + PLL_DESKEW_HSDIV_CTRL(pll->idx, i));
+			ctrl = (uint32_t)readl((uintptr_t)pll->base + (uintptr_t)PLL_DESKEW_HSDIV_CTRL(pll->idx, i));
 			ctrl |= PLL_DESKEW_HSDIV_CTRL_CLKOUT_EN;
-			ti_clk_writel(ctrl, pll->base +
-				      PLL_DESKEW_HSDIV_CTRL(pll->idx, i));
+			ti_clk_writel(ctrl, (uintptr_t)pll->base +
+				      (uintptr_t)PLL_DESKEW_HSDIV_CTRL(pll->idx, i));
 		}
 	}
